Crap, this was supposed to be a double, not a float. Panopticon can't even serialize floats.

+++ projects/mcell/src/java/com/threerings/msoy/aggregators/DailyExchangeRate.java
@@ -44,6 +44,6 @@ public class DailyExchangeRate
         throws IOException
     {
         writer.write(builder.create(
-            "date", key.timestamp, "rate", (float) (result.total / result.count)));
+            "date", key.timestamp, "rate", result.total / result.count));
     }
 }
